styling added

+++ models/asl.js
@@ -1,7 +1,7 @@
 module.exports = function(sequelize, DataTypes) {
   var Asl = sequelize.define("Asl", {
     search: DataTypes.STRING,
-    count: DataTypes.INTEGER
+    count: DataTypes.DECIMAL
   });
